Revert "Refactoring: Move cache features to javaparser-core"
This reverts commit c6a0e51fa1c5c91ec6db8dbd8e737e61ef0f3cf2.
