+++ javaparser-core/pom.xml
@@ -27,13 +27,6 @@
     <properties>
         <java.version>1.8</java.version>
     </properties>
-    
-    <dependencies>
-		<dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-        </dependency>
-	</dependencies>
 
     <build>
         <plugins>
