+++ javaparser-symbol-solver-core/src/main/java/com/github/javaparser/symbolsolver/cache/GuavaCache.java
@@ -19,63 +19,74 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.resolution.cache;
+package com.github.javaparser.symbolsolver.cache;
 
+import java.util.Objects;
 import java.util.Optional;
 
 /**
- * A cache implementation that does not store any information.
+ * This class is used to wrap a Guava {@link com.google.common.cache.Cache}.
  *
- * @param <K> The key type.
- * @param <V> The value type.
+ * @param <K> The type of the key.
+ * @param <V> The type of the value.
  */
-public class NoCache<K, V> implements Cache<K, V> {
+public class GuavaCache<K, V> implements Cache<K, V>  {
 
     /**
-     * Create a new instance.
+     * Wrap a Guava cache with a custom cache.
+     *
+     * @param guavaCache The guava cache to be wrapped-
      *
      * @param <expectedK> The expected type for the key.
      * @param <expectedV> The expected type for the value.
      *
      * @return A newly created instance of {@link NoCache}.
      */
-    public static <expectedK, expectedV> NoCache<expectedK, expectedV> create() {
-        return new NoCache<>();
+    public static <expectedK, expectedV> GuavaCache<expectedK, expectedV> create(com.google.common.cache.Cache<expectedK, expectedV> guavaCache) {
+        return new GuavaCache<>(guavaCache);
+    }
+
+    private final com.google.common.cache.Cache<K, V> guavaCache;
+
+    public GuavaCache(com.google.common.cache.Cache<K, V> guavaCache) {
+        this.guavaCache = Objects.requireNonNull(guavaCache, "The argument GuavaCache can't be null.");
     }
 
     @Override
     public void put(K key, V value) {
-        // Nothing to do here.
+        guavaCache.put(key, value);
     }
 
     @Override
     public Optional<V> get(K key) {
-        return Optional.empty();
+        return Optional.ofNullable(
+                guavaCache.getIfPresent(key)
+        );
     }
 
     @Override
     public void remove(K key) {
-        // Nothing to do here.
+        guavaCache.invalidate(key);
     }
 
     @Override
     public void removeAll() {
-        // Nothing to do here.
+        guavaCache.invalidateAll();
     }
 
     @Override
     public boolean contains(K key) {
-        return false;
+        return get(key).isPresent();
     }
 
     @Override
     public long size() {
-        return 0;
+        return guavaCache.size();
     }
 
     @Override
     public boolean isEmpty() {
-        return true;
+        return size() == 0;
     }
 
 }
